Added specific amp tests for accessing the image in a different way on amp
