+++ src/integration/pages/onDemandRadioEpisode/ampOnlyEpisodeTests.js
@@ -1,0 +1,9 @@
+export default () => {
+  it('I can see the hero image', () => {
+    const imageEl = document.querySelector('main div amp-img img');
+
+    expect(imageEl).toBeInTheDocument();
+    expect(imageEl).toBeTruthy();
+    expect(imageEl.getAttribute('src')).toMatchSnapshot();
+  });
+};
