+++ src/integration/pages/onDemandRadioEpisode/ampTests.js
@@ -1,6 +1,8 @@
 import { runCoreAmpTests, runAmpAnalyticsTests } from '../../common';
+import runAmpOnlyEpisodeTests from './ampOnlyEpisodeTests';
 
 export default () => {
   runCoreAmpTests();
   runAmpAnalyticsTests();
+  runAmpOnlyEpisodeTests();
 };
