+++ src/integration/pages/onDemandRadioEpisode/canonicalTests.js
@@ -2,8 +2,10 @@ import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
 } from '../../common';
+import runCanonicalOnlyEpisodeTests from './canonicalOnlyEpisodeTests';
 
 export default () => {
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
+  runCanonicalOnlyEpisodeTests();
 };
