+++ src/integration/pages/onDemandRadioEpisode/pashto.expired/amp.test.js
@@ -4,9 +4,11 @@
  */
 
 import runExpiredEpisodeTests from '../expiredEpisodeTests';
+import runAmpTests from '../ampTests';
 
 describe('AMP', () => {
   describe(pageType, () => {
     runExpiredEpisodeTests();
+    runAmpTests();
   });
 });
